+++ CHANGELOG.md
@@ -6,7 +6,10 @@
 
 - Added networks and providers configs for Arbitrum Sepolia network.
 - Added arbitrum networks as companion networks of bsc networks.
+
+### Breaking changes
 - Updated to dependency version `@animoca/ethereum-contract-helpers@1.0.3`.
+- Updated to use arbsepolia instead of deprecated arbgoerli as companion networks for testnets.
 
 ## 3.0.3